Create ProcessWindow component
 - Create ProcessWindow Component with generic Windows style preset, Linux and JOptionPane versions
 - Move original JOptionPane component to /window, base on ProcessWindow component, fix imports in components previously using v1 JOP
