+++ src/src/components/Window/JOptionPane/JOptionPane.scss
@@ -8,26 +8,6 @@
   font-family: Arial, Helvetica, sans-serif;
   background: #ececec;
 
-  .toolbar {
-    display: flex;
-    justify-content: space-between;
-    background: #0c0f16;
-    color: white;
-
-    .title, .close {
-      padding: 5px 10px;
-
-      svg {
-        color: white;
-        fill: white;
-      }
-    }
-
-    .close:hover {
-      background: red;
-    }
-  }
-
   .content {
     flex: 100;
     padding: 10px 10px;
