+++ src/src/components/Window/LinuxWindow/LinuxWindow.scss
@@ -1,0 +1,34 @@
+.linux-window {
+  border-radius: 5px 5px 2px 2px;
+
+  .toolbar {
+    flex-direction: row-reverse;
+    justify-content: flex-end;
+    background: #4f4d46;
+    border-bottom: 2px solid black;
+
+    .close {
+      background: #cd5a33;
+      border-radius: 50%;
+      position: relative;
+      margin: 6px;
+      height: 20px;
+      width: 20px;
+
+      &:hover {
+        background: #ec5641;
+      }
+
+      svg {
+        height: 12px;
+        width: 12px;
+        position: absolute;
+        color: black;
+        fill: black;
+        margin: 0 auto;
+        top: 4px;
+        left: 4px;
+      }
+    }
+  }
+}
